components/Factory/requirements.txt
asn1crypto==1.2.0
csrbuilder==0.10.1
pycrypto==2.6.1
pycryptodomex==3.9.4
